:sparkles: Add detail screen (with Compose)

+++ feature/boardgame/src/main/java/fr/boitakub/bogadex/boardgame/model/BoardGameBggStatistic.kt
@@ -5,10 +5,10 @@ import java.util.Date
 
 data class BoardGameBggStatistic(
     @ColumnInfo(name = "num_of_rating_votes") val numOfRatingVotes: Int = 0,
-    @ColumnInfo(name = "average") val average: Double = 0.0,
-    @ColumnInfo(name = "bayes_average") val bayesAverage: Double = 0.0,
+    @ColumnInfo(name = "average") val average: Float = 0f,
+    @ColumnInfo(name = "bayes_average") val bayesAverage: Float = 0f,
     @ColumnInfo(name = "num_of_weight_votes") val numOfWeightVotes: Int = 0,
-    @ColumnInfo(name = "average_weight") val averageWeight: Double = 0.0,
+    @ColumnInfo(name = "average_weight") val averageWeight: Float = 0f,
     @ColumnInfo(name = "num_of_owns") val numOfOwns: Int = 0,
     @ColumnInfo(name = "num_of_wanted") val numOfWanted: Int = 0,
     @ColumnInfo(name = "num_of_wished") val numOfWished: Int = 0,
